--- yolox/data/datasets/custom.py
@@ -1,134 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding:utf-8 -*-
-# Copyright (c) Megvii, Inc. and its affiliates.
-
-import json
-import os
-
-import cv2
-import numpy as np
-from pycocotools.coco import COCO
-
-from ..dataloading import get_yolox_datadir
-from .datasets_wrapper import Dataset
-
-
-class CustomDataset(Dataset):
-    """
-    Custom dataset class.
-    """
-
-    def __init__(
-        self,
-        data_dir=None,
-        json_file="instances_train2017.json",
-        name="train2017",
-        img_size=(416, 416),
-        preproc=None,
-    ):
-        """
-        COCO dataset initialization. Annotation data are read into memory by COCO API.
-        Args:
-            data_dir (str): dataset root directory
-            json_file (str): COCO json file name
-            name (str): COCO data name (e.g. 'train2017' or 'val2017')
-            img_size (int): target image size after pre-processing
-            preproc: data augmentation strategy
-        """
-        super().__init__(img_size)
-        if data_dir is None:
-            data_dir = os.path.join(get_yolox_datadir(), "COCO")
-        self.data_dir = data_dir
-        self.json_file = json_file
-
-        self.coco = COCO(os.path.join(self.data_dir, "annotations",
-                                      self.json_file))
-        self.ids = self.coco.getImgIds()
-        self.class_ids = sorted(self.coco.getCatIds())
-        cats = self.coco.loadCats(self.coco.getCatIds())
-        self._classes = tuple([c["name"] for c in cats])
-        self.name = name
-        self.img_size = img_size
-        self.preproc = preproc
-
-    def __len__(self):
-        return len(self.ids)
-
-    def load_anno(self, index):
-        id_ = self.ids[index]
-        anno_ids = self.coco.getAnnIds(imgIds=[int(id_)], iscrowd=False)
-        annotations = self.coco.loadAnns(anno_ids)
-
-        im_ann = self.coco.loadImgs(id_)[0]
-        width = im_ann["width"]
-        height = im_ann["height"]
-
-        # load labels
-        valid_objs = []
-        for obj in annotations:
-            x1 = np.max((0, obj["bbox"][0]))
-            y1 = np.max((0, obj["bbox"][1]))
-            x2 = np.min((width - 1, x1 + np.max((0, obj["bbox"][2] - 1))))
-            y2 = np.min((height - 1, y1 + np.max((0, obj["bbox"][3] - 1))))
-            if obj["area"] > 0 and x2 >= x1 and y2 >= y1:
-                obj["clean_bbox"] = [x1, y1, x2, y2]
-                valid_objs.append(obj)
-        objs = valid_objs
-        num_objs = len(objs)
-
-        res = np.zeros((num_objs, 5))
-
-        for ix, obj in enumerate(objs):
-            cls = self.class_ids.index(obj["category_id"])
-            res[ix, 0:4] = obj["clean_bbox"]
-            res[ix, 4] = cls
-
-        return res
-
-    def pull_item(self, index):
-        id_ = self.ids[index]
-
-        im_ann = self.coco.loadImgs(id_)[0]
-        width = im_ann["width"]
-        height = im_ann["height"]
-
-        # load image and preprocess
-        img_file = os.path.join(
-            self.data_dir, self.name, "{:012}".format(id_) + ".jpg"
-        )
-        img = cv2.imread(img_file)
-        assert img is not None
-
-        # load anno
-        res = self.load_anno(index)
-        img_info = (height, width)
-
-        return img, res, img_info, id_
-
-    @Dataset.resize_getitem
-    def __getitem__(self, index):
-        """
-        One image / label pair for the given index is picked up and pre-processed.
-
-        Args:
-            index (int): data index
-
-        Returns:
-            img (numpy.ndarray): pre-processed image
-            padded_labels (torch.Tensor): pre-processed label data.
-                The shape is :math:`[max_labels, 5]`.
-                each label consists of [class, xc, yc, w, h]:
-                    class (float): class index.
-                    xc, yc (float) : center of bbox whose values range from 0 to 1.
-                    w, h (float) : size of bbox whose values range from 0 to 1.
-            info_img : tuple of h, w, nh, nw, dx, dy.
-                h, w (int): original shape of the image
-                nh, nw (int): shape of the resized image without padding
-                dx, dy (int): pad size
-            img_id (int): same as the input index. Used for evaluation.
-        """
-        img, res, img_info, img_id = self.pull_item(index)
-
-        if self.preproc is not None:
-            img, target = self.preproc(img, res, self.input_dim)
-        return img, target, img_info, img_id
