--- yolox/data/datasets/custom_classes.py
@@ -1,8 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding:utf-8 -*-
-# Copyright (c) Megvii, Inc. and its affiliates.
-
-CUSTOM_CLASSES = (
-    "cat",
-    "dog",
-)
