+++ yolox/exp/yolox_base.py
@@ -2,17 +2,18 @@
 # -*- coding:utf-8 -*-
 # Copyright (c) 2014-2021 Megvii Inc. All rights reserved.
 
+import os
+import random
+
 import torch
 import torch.distributed as dist
 import torch.nn as nn
-
-import os
-import random
 
 from .base_exp import BaseExp
 
 
 class Exp(BaseExp):
+
     def __init__(self):
         super().__init__()
 
@@ -26,6 +27,7 @@ class Exp(BaseExp):
         self.data_num_workers = 4
         self.input_size = (640, 640)
         self.random_size = (14, 26)
+        self.data_dir = None
         self.train_ann = "instances_train2017.json"
         self.val_ann = "instances_val2017.json"
 
@@ -60,7 +62,7 @@ class Exp(BaseExp):
         self.nmsthre = 0.65
 
     def get_model(self):
-        from yolox.models import YOLOPAFPN, YOLOX, YOLOXHead
+        from yolox.models import YOLOX, YOLOPAFPN, YOLOXHead
 
         def init_yolo(M):
             for m in M.modules():
@@ -81,15 +83,15 @@ class Exp(BaseExp):
     def get_data_loader(self, batch_size, is_distributed, no_aug=False):
         from yolox.data import (
             COCODataset,
+            TrainTransform,
+            YoloBatchSampler,
             DataLoader,
             InfiniteSampler,
             MosaicDetection,
-            TrainTransform,
-            YoloBatchSampler
         )
 
         dataset = COCODataset(
-            data_dir=None,
+            data_dir=self.data_dir,
             json_file=self.train_ann,
             img_size=self.input_size,
             preproc=TrainTransform(
@@ -121,7 +123,9 @@ class Exp(BaseExp):
         if is_distributed:
             batch_size = batch_size // dist.get_world_size()
 
-        sampler = InfiniteSampler(len(self.dataset), seed=self.seed if self.seed else 0)
+        sampler = InfiniteSampler(
+            len(self.dataset), seed=self.seed if self.seed else 0
+        )
 
         batch_sampler = YoloBatchSampler(
             sampler=sampler,
@@ -141,7 +145,7 @@ class Exp(BaseExp):
         tensor = torch.LongTensor(2).cuda()
 
         if rank == 0:
-            size_factor = self.input_size[1] * 1.0 / self.input_size[0]
+            size_factor = self.input_size[1] * 1. / self.input_size[0]
             size = random.randint(*self.random_size)
             size = (int(32 * size), 32 * int(size * size_factor))
             tensor[0] = size[0]
@@ -186,7 +190,6 @@ class Exp(BaseExp):
 
     def get_lr_scheduler(self, lr, iters_per_epoch):
         from yolox.utils import LRScheduler
-
         scheduler = LRScheduler(
             self.scheduler,
             lr,
@@ -203,7 +206,7 @@ class Exp(BaseExp):
         from yolox.data import COCODataset, ValTransform
 
         valdataset = COCODataset(
-            data_dir=None,
+            data_dir=self.data_dir,
             json_file=self.val_ann if not testdev else "image_info_test-dev2017.json",
             name="val2017" if not testdev else "test2017",
             img_size=self.test_size,
